Revert

+++ include/pybind11/detail/python_compat.h
@@ -27,8 +27,8 @@ constexpr auto pybind_vectorcall_call = PyVectorcall_Call;
 
 constexpr size_t pybind_vectorcall_arguments_offset = PY_VECTORCALL_ARGUMENTS_OFFSET;
 
-inline Py_ssize_t pybind_vectorcall_nargs(size_t nargs_with_flag) {
-    return PyVectorcall_NARGS(nargs_with_flag);
+inline size_t pybind_vectorcall_nargs(size_t nargs_with_flag) {
+    return static_cast<size_t>(PyVectorcall_NARGS(nargs_with_flag));
 }
 
 #else
@@ -43,9 +43,7 @@ using pybind_vectorcallfunc = PyObject *(*) (PyObject *, PyObject *const *, size
 constexpr size_t pybind_vectorcall_arguments_offset
     = (static_cast<size_t>(1)) << (static_cast<size_t>(8 * sizeof(size_t) - 1));
 
-inline Py_ssize_t pybind_vectorcall_nargs(size_t n) {
-    return n & ~pybind_vectorcall_arguments_offset;
-}
+inline size_t pybind_vectorcall_nargs(size_t n) { return n & ~pybind_vectorcall_arguments_offset; }
 
 inline PyObject *pybind_vectorcall_call(PyObject *callable, PyObject *args, PyObject *kwargs) {
     Py_ssize_t num_pos_args = PyTuple_GET_SIZE(args);
